requirements.txt
# ─────────────────────────────────────────────
# Podcast Mariano AI — Dependencias
# ─────────────────────────────────────────────

# Orquestación de agentes
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Interfaz web
streamlit>=1.35.0

# Validación de datos
pydantic>=2.0.0

# Audio (ElevenLabs)
requests>=2.31.0

# Variables de entorno
python-dotenv>=1.0.0
